requirements.txt
aiogram
qrcode
pillow
python-dotenv